+++ RELEASE_NOTES.md
@@ -4,6 +4,7 @@
 
 - added: `push=no` and `rebase=no` branch qualifiers that control push and rebase behaviour of `git machete traverse`
 - added: `machete.github.domain` config key to support GitHub Enterprise domains
+- added: support for per-domain entries in `~/.github-token` file
 
 ## New in git-machete 3.13.2